lab5/start/requirements.txt
Flask
Flask-WTF
python-dotenv
